add navigation back to map
